Test 1

+++ src/Database/Database.php
@@ -16,7 +16,8 @@ class Database
     // Simple Types
     const VAR_STRING = 'string';
     const VAR_INTEGER = 'integer';
-    const VAR_FLOAT = 'float';
+    // const VAR_FLOAT = 'float';
+    const VAR_FLOAT = 'double';
     const VAR_BOOLEAN = 'boolean';
 
     // Relationships Types
